refactor: rename main export file
