+++ index.ts
@@ -1,2 +1,2 @@
-export { subagentLogging as default } from '@/src/subagent-logging.ts'
-export { subagentLogging } from '@/src/subagent-logging.ts'
+export { subagentLogging as default } from '@/opencode-subagent-logging'
+export { subagentLogging } from '@/opencode-subagent-logging'
