+++ src/opencode-subagent-logging.ts
